requirements.txt
Flask~=3.0.0
gunicorn~=21.2.0
requests~=2.31.0
Pillow~=10.1.0
psycopg~=3.1.12
python-dotenv~=1.0.0
PyJWT~=2.8.0